classify logical agg logical and receiver func into a unified file
