+++ pkg/planner/core/property_cols_prune.go
@@ -186,25 +186,3 @@ func (p *LogicalJoin) PreparePossibleProperties(_ *expression.Schema, childrenPr
 	}
 	return resultProperties
 }
-
-// PreparePossibleProperties implements base.LogicalPlan PreparePossibleProperties interface.
-func (la *LogicalAggregation) PreparePossibleProperties(_ *expression.Schema, childrenProperties ...[][]*expression.Column) [][]*expression.Column {
-	childProps := childrenProperties[0]
-	// If there's no group-by item, the stream aggregation could have no order property. So we can add an empty property
-	// when its group-by item is empty.
-	if len(la.GroupByItems) == 0 {
-		la.PossibleProperties = [][]*expression.Column{nil}
-		return nil
-	}
-	resultProperties := make([][]*expression.Column, 0, len(childProps))
-	groupByCols := la.GetGroupByCols()
-	for _, possibleChildProperty := range childProps {
-		sortColOffsets := getMaxSortPrefix(possibleChildProperty, groupByCols)
-		if len(sortColOffsets) == len(groupByCols) {
-			prop := possibleChildProperty[:len(groupByCols)]
-			resultProperties = append(resultProperties, prop)
-		}
-	}
-	la.PossibleProperties = resultProperties
-	return resultProperties
-}
